Make auto file generation idompotent
- Improve test coverage

+++ stitcher/stitcher.go
@@ -5,6 +5,7 @@ import (
 	"github.com/spf13/afero"
 	"github.com/vikramyadav1/weaver/parsers"
 	"github.com/vikramyadav1/weaver/renderer"
+	"log"
 	"path/filepath"
 	"strings"
 	"time"
@@ -39,32 +40,56 @@ func (s stitcher) Stitch() error {
 
 func tryCreateModel(s *stitcher) error {
 	dirPath := filepath.Join(s.rootDir, "models", s.rd.Name)
+	fp := filepath.Join(dirPath, "model.go")
+
+	exists, _ := afero.Exists(s.fs, fp)
+	if exists {
+		log.Printf("Skipping model for %s generation as it already exists.\n", s.rd.Name)
+		return nil
+	}
+
 	err := s.fs.MkdirAll(dirPath, 0777)
 	if err != nil {
 		return err
 	}
 
-	return afero.WriteFile(s.fs, filepath.Join(dirPath, "model.go"), s.renderings.Model(), 0777)
+	return afero.WriteFile(s.fs, fp, s.renderings.Model(), 0777)
 }
 
 func tryCreateRepository(s *stitcher) error {
 	dirPath := filepath.Join(s.rootDir, "models", s.rd.Name)
+	fp := filepath.Join(dirPath, "repository.go")
+
+	exists, _ := afero.Exists(s.fs, fp)
+	if exists {
+		log.Printf("Skipping model for %s generation as it already exists.\n", s.rd.Name)
+		return nil
+	}
+
 	err := s.fs.MkdirAll(dirPath, 0777)
 	if err != nil {
 		return err
 	}
 
-	return afero.WriteFile(s.fs, filepath.Join(dirPath, "repository.go"), s.renderings.Repository(), 0777)
+	return afero.WriteFile(s.fs, fp, s.renderings.Repository(), 0777)
 }
 
 func tryCreateServer(s *stitcher) error {
 	dirPath := filepath.Join(s.rootDir, "api")
+	fp := filepath.Join(dirPath, s.rd.Name+"Server.go")
+
+	exists, _ := afero.Exists(s.fs, fp)
+	if exists {
+		log.Printf("Skipping model for %s generation as it already exists.\n", s.rd.Name)
+		return nil
+	}
+
 	err := s.fs.MkdirAll(dirPath, 0777)
 	if err != nil {
 		return err
 	}
 
-	return afero.WriteFile(s.fs, filepath.Join(dirPath, s.rd.Name+"Server.go"), s.renderings.Server(), 0777)
+	return afero.WriteFile(s.fs, fp, s.renderings.Server(), 0777)
 }
 
 func tryCreateMain(s *stitcher) error {
@@ -80,7 +105,9 @@ func tryCreateMain(s *stitcher) error {
 
 	mainRendering = s.renderings.PartialMain()
 	mainContents, err := afero.ReadFile(s.fs, mainFilepath)
-	fmt.Printf("Error reading main file.\nError: %v", err)
+	if err != nil {
+		fmt.Printf("Error reading main file.\nError: %v", err)
+	}
 	newMainContent := strings.Replace(string(mainContents), "//weaver:renderEnd", string(mainRendering), 1)
 
 	return afero.WriteFile(s.fs, filepath.Join(s.rootDir, "main.go"), []byte(newMainContent), 0777)
@@ -88,7 +115,17 @@ func tryCreateMain(s *stitcher) error {
 
 func tryCreateMigrations(s *stitcher) error {
 	dirPath := filepath.Join(s.rootDir, "migrations")
-	err := s.fs.MkdirAll(dirPath, 0777)
+	existingUpMigrations, err := afero.Glob(s.fs, filepath.Join(dirPath, "*.up.sql"))
+	if err != nil {
+		log.Fatalf("Error searching for up migraion. error: %v\n", err)
+	}
+
+	existingDownMigrations, err := afero.Glob(s.fs, filepath.Join(dirPath, "*.up.sql"))
+	if err != nil {
+		log.Fatalf("Error searching for down migraion. error: %v\n", err)
+	}
+
+	err = s.fs.MkdirAll(dirPath, 0777)
 	if err != nil {
 		return err
 	}
@@ -101,10 +138,16 @@ func tryCreateMigrations(s *stitcher) error {
 	upFilename := fmt.Sprintf("%s.up.sql", formatted)
 	downFilename := fmt.Sprintf("%s.down.sql", formatted)
 
-	err = afero.WriteFile(s.fs, filepath.Join(dirPath, upFilename), s.renderings.UpMigration(), 0777)
-	if err != nil {
-		return err
+	if len(existingUpMigrations) == 0 {
+		err = afero.WriteFile(s.fs, filepath.Join(dirPath, upFilename), s.renderings.UpMigration(), 0777)
+		if err != nil {
+			return err
+		}
 	}
 
-	return afero.WriteFile(s.fs, filepath.Join(dirPath, downFilename), s.renderings.DownMigration(), 0777)
+	if len(existingDownMigrations) == 0 {
+		return afero.WriteFile(s.fs, filepath.Join(dirPath, downFilename), s.renderings.DownMigration(), 0777)
+	}
+
+	return nil
 }
